add some first convenience functions to work with Nlopt
Adds some first simple convenience functions, to work with Nlopt,
which take care of transforming the Nim types to compatible types,
deals with Nlopt's expected pointers etc.

Provides some test on an example, in which we optimize the rotation
angle / the eccentricity of a cluster of points (although we use a
line of data points from (0, 0) -> (99, 99) in this case, because I'm
too lazy to create a more interesting cluster right now...).