allow for multidimensional problems with `NloptOpt` obj incl. bounds
Finally allows for the creation of optimizers for more than 1
dimension using the `NloptOpt` object instead of the C bindings
directly. For that also changes the `bound` fields to be a `seq` of
tuples for lower and upper bound of each dimension's parameter.

Also some renaming from snake_case to camelCase.