add possibility to use gradient based algorithms with new interface
Adds the ability to use gradient based algorithms with the new
interface, in which the user only needs to define the main
optimization function. This is then wrapped by the library and a
wrapped function is handed to the C library.

Implemented by a variant object `VarStruct`, which either has a
`FuncProto` or a `FuncProtoGrad` field, which are procs with different
signatures.

In case of the gradient based user function, the user needs to
calculate the gradient in the function and return a tuple of:
(function return value, new gradients)