change signature of `setFunction` to clarify meaning
From a fully generic function with type `T` for the
`vStruct` (previously just `fobj`) argument, we go to a generic
`VarStruct[T]` object. This allows better undertanding of how to use
the `setFunction` and `addInequalityConstraints` procs.

In the current implementation it does not constrain the usability of
the library, since the `VarStruct` object needs to be used
anyways (otherwise we cannot extract the user function from the
`func_data` field in the `optimizeImpl` function). The alternative
would be that the user is forced to use an object, which conforms to
the field names of `VarStruct`. This is redundant, because the user
can already hand arbitrary data to the function, using the `data`
field of `VarStruct`.