fix `FuncProtoGrad` signature and implementation, add example to README
The previous implementation first of all didn't check whether the
given `gradPtr` was actually non nil. In fact even for gradient based
algorithms it sometimes may be nil. So take that into account.

Also, more importantly, the signature of the gradient based functions
is changed so that they do not receive a `grad` argument anymore. That
never actually made sense. The function is supposed to calculate the
gradient. But since we return a `seq[float]` as a second tuple
element, there's no need for a `grad` argument, which may be used to
change the gradient in place (if it was a `var` argument).