add `userData` field to keep reference to user defined `VarStruct`
For this `NloptOpt` must become generic. This has the downside that
the type of the `VarStruct` must already be given to the `newNloptOpt`
call, despite potentially not having defined the `VarStruct` yet.

So a call to `newNloptOpt` mus explicitly give the type `T`.

A better solution would be to remove the convention of having
to call `setFunction` explicitly and rather give the proc to be
optimized and the `VarStruct` type to the `newNloptOpt` proc.